requirements.txt
python-telegram-bot==21.6
openai==1.51.0
supabase==2.9.0
python-dotenv==1.0.1
